Fix hiding the window title in macOS
In f5a57a3c7578dfabd560c5ecb26e465242940521 I introduced a bug where only the window title of the first OS window would be hidden. This commit fixes that bug.
The bug was caused by the code hiding the window title being in an `if` statement that only executes once.