Fix flickering of prompt during window resize
Works by keeping the old prompt unreflowed rather than clearing it.
There may still be some flicker for people using long or right side
prompts, but that cant be avoided, since we cannot know how the shell
will redraw after the resize. But in the common case of a left side
smallish prompt that fits in the resized window, the flicker is
eliminated.

It means we have to do some more copying work on resize, but the nicer
visuals is worth it, IMO.