shell-integration/ssh: Support arch=arm64, for MacOS
On some Macs (e.g. M1, 2020), `uname` returns 'Darwin' and `uname -m`
retuns 'arm64', this case is not handled in shell-integration. As a
result, ssh kitten transfer fails with:

>> Unknown CPU architecture arm64

Fix this by adding the case for arch=arm64 in kitten and kitty
shell-integration scripts.

Tested this on Mac Pro (M1, 2020).