Wayland: Move creation of OpenGL context to end of window creation function
Apparently with explicit sync it is now not possible to attach a
temporary buffer to a surface that also has an OpenGL context. So we
first attach the temporary buffer, and only after we are done waiting
for window creation do we attach the OpenGL context to the window
surface.

Fixes #7767 (maybe, not installing sway-git to check)